use 64 bit integer

+++ src/main/java/com/brinkcommerce/api/management/order/delivery/model/response/BrinkDeliveryPostResponse.java
@@ -29,10 +29,10 @@ public record BrinkDeliveryPostResponse(
         Instant updated,
         Instant started,
         Instant restarted,
-        Integer restarts,
+        Long restarts,
         Instant completed,
         Instant failed,
-        Integer revision
+        Long revision
 ) {}
